Add register_addon method
Registering addons is currently done via setting an environment variable
for the endpoint and using kensa. The logic is also implemented in some
of our components to register addons on the bootstrap process.
cloud&heat use a concatenation of tenant_name and email for the login,
this breaks the kensa tool.

This commit simplifies the register process in our components and enables
us to add the register addon functionality in the cctrl/exo/dotcloudng/cnh.