Handle SSH Public Key auth
Do a separation between basic auth and ssh auth. Former will remain
the functionality from `create_token` method (this will act now as a
wrapper of it), the latter will handle the ssh token auth with server
side via a "ccssh" header.

Implement a `create_ssh_token` public method that will request a
ssh token to server side which will be associated with the user
itself.

Create `APIException` class that will serve as base class for all of the
cclib custom exceptions.

Cleanup Request class, so auth details are not explicitly
defined on the constructor of an object of this kind.