Make Request class private
Previously, Request was a public class available within the cclib
module, although it was stated that it's only used internally
by the API class itself.

This commit attempts to make this statement consistent, so we don't
only discourage its public use, but also rename it as _Request,
following python naming conventions.